fix : securityConfig 파일 위치 api 모듈로 위치를 변경 infra 모듈은 외부 시스템(redis,db) 담당 모듈이고 securityConfig 는 API 요청 처리(인증,인가) 이기 때문에 api 모둘에 배치되는게 맞다고 판단
